reverted to older style DBAccess seperation on Cheawon's suggestion
