+++ src/main/java/database/DBAccess.java
@@ -4,6 +4,7 @@ import com.google.gson.Gson;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
+import database.converters.JsonConverter;
 
 import java.io.IOException;
 import java.util.ArrayList;
