+++ src/main/java/database/RoomDBAccess.java
@@ -1,0 +1,52 @@
+package database;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+
+import database.converters.RoomConverter;
+import host_room.HostRoomDBGateway;
+import join_room.JoinByIDDBGateway;
+import leave_room.LeaveRoomDBGateway;
+import messaging.MessageDBGateway;
+
+public class RoomDBAccess extends DBAccess<RoomDBModel> implements HostRoomDBGateway, JoinByIDDBGateway, LeaveRoomDBGateway, MessageDBGateway {
+
+    public RoomDBAccess(HttpClient httpClient, RoomConverter converter) {
+        super(httpClient, converter);
+    }
+
+    @Override
+    public List<RoomDBModel> getRooms() {
+        try {
+            return getRows();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        return new ArrayList<>();
+    }
+
+    @Override
+    public RoomDBModel getARoom(Integer roomID) {
+        try {
+            return getARow(roomID);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+    @Override
+    public void updateARoom(RoomDBModel request) {
+        try {
+            updateARow(request.getRoomID(), request);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    @Override
+    protected String getRoute() {
+        return "rooms";
+    }
+}
