+++ src/main/java/database/converters/JsonConverter.java
@@ -1,4 +1,4 @@
-package database;
+package database.converters;
 
 import com.google.gson.JsonObject;
