+++ src/main/java/database/converters/RoomConverter.java
@@ -1,9 +1,10 @@
-package database;
+package database.converters;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
+import database.RoomDBModel;
 
 import java.util.ArrayList;
 import java.util.List;
