+++ src/main/java/host_room/HostRoomDBGateway.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface HostRoomDBGateway {
     List<RoomDBModel> getRooms();
-    void hostARoom(RoomDBModel request);
+    void updateARoom(RoomDBModel request);
 }
