+++ src/main/java/join_room/JoinByIDDBGateway.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface JoinByIDDBGateway {
     List<RoomDBModel> getRooms();
-    void joinARoom(RoomDBModel request);
+    void updateARoom(RoomDBModel request);
 }
