+++ src/main/java/join_room/JoinByIDInteractor.java
@@ -23,7 +23,7 @@ public class JoinByIDInteractor implements JoinByIDInputBoundary {
                 List<Integer> activeUsers = room.getActiveUserIDs();
                 activeUsers.add(User.getUserID());
                 room.setActiveUserIDs(activeUsers);
-                roomDBGateway.joinARoom(room);
+                roomDBGateway.updateARoom(room);
                 Room.setRoom(room.getRoomID(), room.getRoomName(),
                         room.getHostID(), room.getActiveUserIDs(), room.getMessageHistory());
                 presenter.prepareRoomView(Room.getMessageHistory());
