+++ src/main/java/leave_room/LeaveRoomInteractor.java
@@ -22,7 +22,7 @@ public class LeaveRoomInteractor implements LeaveRoomInputBoundary {
         if (activeUsers.contains(User.getUserID())) {
             activeUsers.remove(User.getUserID());
             room.setActiveUserIDs(activeUsers);
-            roomDBGateway.leaveARoom(room);
+            roomDBGateway.updateARoom(room);
             leaveRoomOutputBoundary.prepareJoinOrHostView();
             return;
         }
