+++ src/main/java/messaging/MessageDBGateway.java
@@ -4,5 +4,5 @@ import database.RoomDBModel;
 
 public interface MessageDBGateway {
     RoomDBModel getARoom(Integer roomID);
-    void sendAMessage(RoomDBModel room);
+    void updateARoom(RoomDBModel room);
 }
